fix: combine get shared album api as get album api
